+++ modules/manager/controller_setup.go
@@ -92,8 +92,10 @@ const (
 	KongCACertificateControllerName = "KongCACertificate"
 	// KongCertificateControllerName is the name of the KongCertificate controller.
 	KongCertificateControllerName = "KongCertificate"
-	// KongVaultContollerName is the name of KongVault controller.
+	// KongVaultControllerName is the name of KongVault controller.
 	KongVaultControllerName = "KongVault"
+	// KongKeyControllerName is the name of KongKey controller.
+	KongKeyControllerName = "KongKey"
 )
 
 // SetupControllersShim runs SetupControllers and returns its result as a slice of the map values.
@@ -452,6 +454,15 @@ func SetupControllers(mgr manager.Manager, c *Config) (map[string]ControllerDef,
 					c.DevelopmentMode,
 					mgr.GetClient(),
 					konnect.WithKonnectEntitySyncPeriod[configurationv1alpha1.KongCredentialAPIKey](c.KonnectSyncPeriod),
+				),
+			},
+			KongKeyControllerName: {
+				Enabled: c.KonnectControllersEnabled,
+				Controller: konnect.NewKonnectEntityReconciler(
+					sdkFactory,
+					c.DevelopmentMode,
+					mgr.GetClient(),
+					konnect.WithKonnectEntitySyncPeriod[configurationv1alpha1.KongKey](c.KonnectSyncPeriod),
 				),
 			},
 			KongPluginControllerName: {